small changes to rasterize, fixed broken website link -- Molly
git-svn-id: https://giordano.mit.edu/nvs/mangle/branches/dev_64bit@6 9f78a35e-8a50-488b-8db1-b351ce30c53d

+++ src/balkanize.c
@@ -162,7 +162,7 @@ int balkanize(int npoly, polygon *poly[/*npoly*/], int npolys, polygon *polys[/*
   int *start;
   int *total;
   int begin, end, p, max_pixel;
-  long double tol;
+  double tol;
 
   poly_sort(npoly, poly, 'p');
